backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.0.0
scikit-learn==1.2.2
numpy==1.26.0
aiohttp==3.8.5
tqdm==4.66.1
torch
torchvision
